// main/types/register-schema.ts
import * as z from "zod";

export const registerSchema = z.object({
    email: z.string().email("Invalid email address"),
    password: z.string().min(8, "Password must be at least 8 characters long"),
    name: z.string().min(6, "Name must be at least 6 characters long"),
})

export type RegisterSchemaValues = z.infer<typeof registerSchema>;